Properly set the IP address in the request URL
So, this module was not working as expected. It works by first calling
`url.parse()` on the given URL. Then it does a DNS resolution on the
hostname, and resets the `hostname` property with the IP address.

Unfortunately, the `url.format()` function prefers the `host` property
over the `hostname` property, so setting `hostname` ends up being a no-op.

For example:

```
> process.version
'v12.18.4'

> const parsed = url.parse('https://example.com/hi')

// Reset `hostname`
> parsed.hostname = '1.2.3.4'
'1.2.3.4'

// `host` is preferred over `hostname`, so `url.format()` returns original URL
> url.format(parsed)
'https://example.com/hi'

> url.format(parsed) === parsed.href
true
```

This PR addresses that issue by overwriting both `host` and `hostname`
appropriately. Ensuring that the proper "Host" request header is set
during redirects is also addressed here, with additional tests.